Eliminate conflict between EscrowBot and SwapBot
Since EscrowBot and SwapBot use the same Reddit account on r/Cash4Cash (u/C4C_Bot), this means that, occasionally, the two scripts both reply to a message. A message that contains a valid EscrowBot command will be responded to by SwapBot, saying no username was provided.

Since all EscrowBot commands all begin with an exclamation point, this can be solved by simply checking if the first char of the message is an exclaimation point, and skipping it if yes.